deskapk/requirements.txt
opencv-python
pyzbar
qrcode[pil]
pillow
pandas
openpyxl